requirements.txt
audioread==3.0.1
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
pandas==2.2.2
pillow==10.3.0
pooch==1.8.1
protobuf==4.25.3
pycparser==2.22
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==2.2.0
setproctitle==1.3.3
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.5.0
torch==2.2.0
tqdm==4.66.4
triton==2.2.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0


